devops/airflow-local-runner/requirements.txt
watchtower
apache-airflow-providers-slack
apache-airflow-providers-amazon
autoflake
boto3
botocore
numpy
pandas
pyarrow
pytz
requests
s3transfer
slack_sdk
tzdata
urllib3
wheel
apache-airflow
apache-airflow-providers-common-sql
apache-airflow-providers-ftp
apache-airflow-providers-http
apache-airflow-providers-imap
apache-airflow-providers-sqlite
apache-airflow-providers-postgres
awswrangler
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-sqlalchemy
clickhouse-connect
